added apdu evm
