+++ app/src/apdu_handler.c
@@ -23,6 +23,7 @@
 
 #include "actions.h"
 #include "addr.h"
+#include "apdu_handler_evm.h"
 #include "app_main.h"
 #include "coin.h"
 #include "coin_evm.h"
@@ -57,36 +58,6 @@ void extract_fil_path(uint32_t rx, uint32_t offset) {
     if (!mainnet && !testnet) {
         THROW(APDU_CODE_DATA_INVALID);
     }
-}
-
-void extract_eth_path(uint32_t rx, uint32_t offset) {
-    tx_initialized = false;
-
-    uint32_t path_len = *(G_io_apdu_buffer + offset);
-
-    if (path_len > MAX_BIP32_PATH || path_len < 1) THROW(APDU_CODE_WRONG_LENGTH);
-
-    if ((rx - offset - 1) < sizeof(uint32_t) * path_len) {
-        THROW(APDU_CODE_WRONG_LENGTH);
-    }
-
-    // first byte at OFFSET_DATA is the path len, so we skip this
-    uint8_t *path_data = G_io_apdu_buffer + offset + 1;
-
-    // hw-app-eth serializes path as BE numbers
-    for (uint8_t i = 0; i < path_len; i++) {
-        hdPathEth[i] = U4BE(path_data, 0);
-        path_data += sizeof(uint32_t);
-    }
-
-    const bool mainnet = hdPathEth[0] == HDPATH_ETH_0_DEFAULT && hdPathEth[1] == HDPATH_ETH_1_DEFAULT;
-
-    if (!mainnet) {
-        THROW(APDU_CODE_DATA_INVALID);
-    }
-
-    // set the hdPath len
-    hdPathEth_len = path_len;
 }
 
 __Z_INLINE bool process_chunk(__Z_UNUSED volatile uint32_t *tx, uint32_t rx) {
@@ -189,104 +160,6 @@ __Z_INLINE bool process_rawbytes_chunk(__Z_UNUSED volatile uint32_t *tx, uint32_
     return false;
 }
 
-__Z_INLINE bool process_chunk_eth(__Z_UNUSED volatile uint32_t *tx, uint32_t rx) {
-    const uint8_t payloadType = G_io_apdu_buffer[OFFSET_PAYLOAD_TYPE];
-
-    if (G_io_apdu_buffer[OFFSET_P2] != 0) {
-        THROW(APDU_CODE_INVALIDP1P2);
-    }
-
-    if (rx < OFFSET_DATA) {
-        THROW(APDU_CODE_WRONG_LENGTH);
-    }
-
-    uint64_t read = 0;
-    uint64_t to_read = 0;
-    uint64_t max_len = 0;
-
-    uint8_t *data = &(G_io_apdu_buffer[OFFSET_DATA]);
-    uint32_t len = rx - OFFSET_DATA;
-
-    uint64_t added;
-    switch (payloadType) {
-        case P1_ETH_FIRST:
-            tx_initialize();
-            tx_reset();
-            extract_eth_path(rx, OFFSET_DATA);
-            // there is not warranties that the first chunk
-            // contains the serialized path only;
-            // so we need to offset the data to point to the first transaction
-            // byte
-            uint32_t path_len = sizeof(uint32_t) * hdPathEth_len;
-
-            // plus the first offset data containing the path len
-            data += path_len + 1;
-            if (len < path_len) {
-                THROW(APDU_CODE_WRONG_LENGTH);
-            }
-
-            // now process the chunk
-            len -= path_len + 1;
-            if (get_tx_rlp_len(data, len, &read, &to_read) != rlp_ok) {
-                THROW(APDU_CODE_DATA_INVALID);
-            }
-
-            // get remaining data len
-            max_len = saturating_add(read, to_read);
-            max_len = MIN(max_len, len);
-
-            added = tx_append(data, max_len);
-            if (added != max_len) {
-                THROW(APDU_CODE_OUTPUT_BUFFER_TOO_SMALL);
-            }
-
-            tx_initialized = true;
-
-            // if the number of bytes read and the number of bytes to read
-            //  is the same as what we read...
-            if ((saturating_add(read, to_read) - len) == 0) {
-                return true;
-            }
-            return false;
-        case P1_ETH_MORE:
-            if (!tx_initialized) {
-                THROW(APDU_CODE_TX_NOT_INITIALIZED);
-            }
-
-            uint64_t buff_len = tx_get_buffer_length();
-            uint8_t *buff_data = tx_get_buffer();
-
-            if (get_tx_rlp_len(buff_data, buff_len, &read, &to_read) != rlp_ok) {
-                THROW(APDU_CODE_DATA_INVALID);
-            }
-
-            uint64_t rlp_read = buff_len - read;
-
-            // either the entire buffer of the remaining bytes we expect
-            uint64_t missing = to_read - rlp_read;
-            max_len = len;
-
-            if (missing < len) max_len = missing;
-
-            added = tx_append(data, max_len);
-
-            if (added != max_len) {
-                tx_initialized = false;
-                THROW(APDU_CODE_OUTPUT_BUFFER_TOO_SMALL);
-            }
-
-            // check if this chunk was the last one
-            if (missing - len == 0) {
-                tx_initialized = false;
-                return true;
-            }
-
-            return false;
-    }
-    THROW(APDU_CODE_INVALIDP1P2);
-    return false;
-}
-
 __Z_INLINE void handleGetAddr(volatile uint32_t *flags, volatile uint32_t *tx, uint32_t rx) {
     extract_fil_path(rx, OFFSET_DATA);
 
@@ -299,31 +172,6 @@ __Z_INLINE void handleGetAddr(volatile uint32_t *flags, volatile uint32_t *tx, u
     }
     if (requireConfirmation) {
         view_review_init(addr_getItem, addr_getNumItems, app_reply_address);
-        view_review_show(REVIEW_ADDRESS);
-        *flags |= IO_ASYNCH_REPLY;
-        return;
-    }
-    *tx = action_addrResponseLen;
-    THROW(APDU_CODE_OK);
-}
-
-__Z_INLINE void handleGetAddrEth(volatile uint32_t *flags, volatile uint32_t *tx, uint32_t rx) {
-    extract_eth_path(rx, OFFSET_DATA);
-
-    uint8_t requireConfirmation = G_io_apdu_buffer[OFFSET_P1];
-    uint8_t with_code = G_io_apdu_buffer[OFFSET_P2];
-
-    if (with_code != P2_CHAINCODE && with_code != P2_NO_CHAINCODE) THROW(APDU_CODE_INVALIDP1P2);
-
-    fil_chain_code = with_code;
-
-    zxerr_t zxerr = app_fill_eth_address();
-    if (zxerr != zxerr_ok) {
-        *tx = 0;
-        THROW(APDU_CODE_DATA_INVALID);
-    }
-    if (requireConfirmation) {
-        view_review_init(eth_addr_getItem, eth_addr_getNumItems, app_reply_address);
         view_review_show(REVIEW_ADDRESS);
         *flags |= IO_ASYNCH_REPLY;
         return;
@@ -396,7 +244,7 @@ __Z_INLINE void handleSignRawBytes(volatile uint32_t *flags, volatile uint32_t *
         const int error_msg_length = strnlen(error_msg, sizeof(G_io_apdu_buffer));
         MEMCPY(G_io_apdu_buffer, error_msg, error_msg_length);
         *tx += (error_msg_length);
-        if (error_code == parser_blindsign_required) {
+        if (error_code == parser_blindsign_mode_required) {
             *flags |= IO_ASYNCH_REPLY;
             view_blindsign_error_show();
         }
@@ -405,37 +253,6 @@ __Z_INLINE void handleSignRawBytes(volatile uint32_t *flags, volatile uint32_t *
 
     CHECK_APP_CANARY()
     view_review_init(tx_getItem, tx_getNumItems, app_sign);
-    view_review_show(REVIEW_TXN);
-    *flags |= IO_ASYNCH_REPLY;
-}
-
-__Z_INLINE void handleSignEth(volatile uint32_t *flags, volatile uint32_t *tx, uint32_t rx) {
-    zemu_log_stack("handleSignEth");
-    if (!process_chunk_eth(tx, rx)) {
-        THROW(APDU_CODE_OK);
-    }
-
-    tx_context_eth();
-
-    CHECK_APP_CANARY()
-
-    uint8_t error_code;
-    const char *error_msg = tx_parse(&error_code);
-    CHECK_APP_CANARY()
-
-    if (error_msg != NULL) {
-        const int error_msg_length = strnlen(error_msg, sizeof(G_io_apdu_buffer));
-        MEMCPY(G_io_apdu_buffer, error_msg, error_msg_length);
-        *tx += (error_msg_length);
-        if (error_code == parser_blindsign_required) {
-            *flags |= IO_ASYNCH_REPLY;
-            view_blindsign_error_show();
-        }
-        THROW(APDU_CODE_DATA_INVALID);
-    }
-
-    CHECK_APP_CANARY()
-    view_review_init(tx_getItem, tx_getNumItems, app_sign_eth);
     view_review_show(REVIEW_TXN);
     *flags |= IO_ASYNCH_REPLY;
 }
