+++ app/src/common/parser_common.h
@@ -69,7 +69,7 @@ typedef enum {
     parser_invalid_prefix,
     // Customs
     parser_expert_mode_required,
-    parser_blindsign_required,
+    parser_blindsign_mode_required,
 } parser_error_t;
 
 // Define the three types
